+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+**Release 0.3**
+
+    * Golems!
+        * Implemented golem seal connector item as an upgrade component
+        * LUA API functions to read information from seals
+
 **Release 0.2**
 
     * First public release